pyproject.toml
[tool.poetry]
name = "fastapi-farm-stack"
version = "0.1.0"
description = ""
authors = ["Yuya Shintani <yuya.shintani.uplus26e7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.70.0"
uvicorn = "0.14.0"
fastapi-csrf-protect = "0.2.1"
motor = "2.5.1"
PyJWT = "2.3.0"
passlib = "1.7.4"
python-decouple = "3.3"
dnspython = "2.1.0"
gunicorn = "20.1.0"
bcrypt = "3.2.0"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
